Split the columnar modules into bespoke files
No changes in functionality expected.
