+++ src/timely-util/src/lib.rs
@@ -19,6 +19,7 @@ pub mod activator;
 pub mod antichain;
 pub mod builder_async;
 pub mod capture;
+pub mod columnar;
 pub mod containers;
 pub mod operator;
 pub mod order;
